rename glue.rs to client.rs
