+++ client/noise-kv/src/client.rs
