+++ client/noise-kv/src/lib.rs
@@ -1,9 +1,9 @@
 #![feature(async_closure)]
 
+pub mod client;
 pub mod contacts;
 pub mod data;
 pub mod devices;
-pub mod glue;
 pub mod groups;
 pub mod permissions;
 pub mod sharing;
